Refactor quote escaping in SpeciesListUpload
Main goal was to avoid modifying strings to escape quotes until the values were about to be inserted into the DB. 

When possible, I used prepared statements to avoid escaping in the first place, and used enquoteLiteral when it was simpler to create the statement dynamically.

saveTaxon uses prepared statements

updateTaxon uses enquoteLiteral

Also asserted types for the Hashtables, so IntelliJ complains a little less.